now sells page support double type quantity

+++ lib/models/sell_model.dart
@@ -1,6 +1,6 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
-
 import 'package:cloud_firestore/cloud_firestore.dart';
+
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 
 class Sell {
   final String id;
@@ -8,7 +8,7 @@ class Sell {
   final String fishName;
   final DateTime date;
   final int price;
-  final int quantity;
+  final double quantity;
   final bool smallFish;
   Sell({
     required this.id,
@@ -41,7 +41,7 @@ class Sell {
       fishName: map['fishName'] as String,
       date: timeStamp.toDate(),
       price: map['price'] as int,
-      quantity: map['quantity'] as int,
+      quantity: map['quantity'] as double,
       smallFish: map['smallFish'] as bool,
     );
   }
